+++ test/fixtures/image.rb
@@ -12,3 +12,12 @@ end
 def fixture_generate_colour
   100.times.map { RAYWHITE }
 end
+
+def fixture_image_from_image
+  t = WHITE
+  o = Colour.new(218, 154, 37, 255)
+  [
+    o, t,
+    t, o,
+  ]
+end